share User type instead of UserState, separate reducer setup from general user utils to address bundle issues
