+++ support-frontend/assets/helpers/redux/subscriptionsStore.ts
@@ -7,7 +7,7 @@ import type { SubscriptionProduct } from 'helpers/productPrice/subscriptions';
 import { renderError } from 'helpers/rendering/render';
 import { createReducer } from 'helpers/subscriptionsForms/subscriptionCheckoutReducer';
 import type { DateYMDString } from 'helpers/types/DateString';
-import { setUpUserState } from 'helpers/user/user';
+import { setUpUserState } from 'helpers/user/reduxSetup';
 import { addAddressSideEffects } from './checkout/address/subscriptionsSideEffects';
 import { addPaymentsSideEffects } from './checkout/payment/subscriptionsSideEffects';
 import { addPersonalDetailsSideEffects } from './checkout/personalDetails/subscriptionsSideEffects';
